Implement BaseEntity
