--- wetube-domain/src/main/comment/Comment.kt
@@ -1,15 +1,0 @@
-import javax.persistence.Entity
-import javax.persistence.ManyToOne
-import java.time.OffsetDateTime
-
-@Entity
-class Comment (
-        @ManyToOne(fetch = FetchType.LAZY)
-        val video: Video,
-        @ManyToOne
-        val user: User,
-        val content: String,
-        val createdAt: OffsetDateTime = OffsetDateTime.now(),
-        @Id @GeneratedValue(strategy = GenerationType.AUTO)
-        val id: Long? = null
-)
