+++ wetube-domain/src/main/kotlin/com/leeparkim/wetube/domain/like/VideoLike.kt
@@ -1,0 +1,15 @@
+package com.leeparkim.wetube.domain.like
+
+import com.leeparkim.wetube.domain.user.User
+import com.leeparkim.wetube.domain.video.Video
+import javax.persistence.Entity
+import javax.persistence.FetchType
+import javax.persistence.ManyToOne
+
+@Entity
+class VideoLike(
+    @ManyToOne(fetch = FetchType.LAZY)
+    val user: User,
+    @ManyToOne(fetch = FetchType.LAZY)
+    val video: Video,
+)
