+++ wetube-domain/src/main/kotlin/com/leeparkim/wetube/domain/user/SocialType.kt
@@ -1,0 +1,5 @@
+package com.leeparkim.wetube.domain.user
+
+enum class SocialType {
+    GOOGLE
+}
