--- wetube-domain/src/main/user/SocialType.kt
@@ -1,3 +1,0 @@
-enum class SocialType {
-    GOOGLE
-}
